+++ src/components/Header.js
@@ -32,4 +32,4 @@ const Header = () => {
       );
 }
 
-export default Header
+export default Header;
